backend/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
flask>=2.0.0
flask-cors>=3.0.0
Pillow>=8.0.0
pdf2image>=1.16.0
requests>=2.25.0
cloudscraper>=1.0.0
numpy>=1.21.0